+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Add `Tucan.Export` wrapper aroung `VegaLite.Export`
+- Add `Tucan.Export` wrapper around `VegaLite.Export`
 
 ## [v0.3.1](https://github.com/pnezis/tucan/tree/v0.3.1) (2024-01-20)